Safari-611.1.10/LayoutTests/fast/css/implicit-property-restore-expected.txt

This test verifies the property implicit flag is reset after adding any css property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element1.style.isPropertyImplicit('padding-top') is false
PASS removedProperties.length is 1
PASS removedProperties[0] is "list-style"
PASS successfullyParsed is true

TEST COMPLETE
